requirements.txt
streamlit
torch
transformers>=4.45.0
peft
accelerate
Pillow
qwen-vl-utils
